+++ Scripts/AdsManager/IAdsNetworkHelper.cs
@@ -15,6 +15,7 @@ public interface IAdsNetworkHelper
     void RequestInterstitialRewardedNoShow(AdPlacement.Type placementType, RewardDelegate onAdLoaded = null);
     void ShowInterstitialRewarded(AdPlacement.Type placementType, RewardDelegate onAdClosed);
     void Reward(AdPlacement.Type placementType, RewardDelegate onFinish);
+    void RequestRewardAd(AdPlacement.Type placementType, RewardDelegate onFinish);
     void RequestAppOpenAd(AdPlacement.Type placementType, RewardDelegate onAdLoaded = null);
     void ShowAppOpenAd(AdPlacement.Type placementType, AdsManager.InterstitialDelegate onAdClosed = null);
 }